code/chatbot/requirements.txt
fastapi
uvicorn
python-dotenv
openai
websockets
prompt_toolkit
requests